requirements.txt
flask
opencv-python
numpy
onnxruntime
insightface
pillow
scikit-learn
scikit-image
scipy
tqdm
requests
matplotlib
albumentations
prettytable
faiss-cpu
